Refactor User class and removed AccountUser class
